feat: add verify-mdx to install pages verification
